Disable ownership check.
Under 5 node membership, the first four nodes own 25% of the ring after the join of node 5, but node 5 ends up with 0% of the ring.  This leads owners_according_to to only return the first four nodes as owners, since node 5 doesn't have any ownership.

This test fails on both the riak_core, fifo-0.9.2 branch and the partisan-enabled fork of Riak Core.